requirements.txt
PySide6
pillow